Add error messages

+++ src/main/java/org/example/javabackend1/Customer/CustomerService.java
@@ -29,6 +29,10 @@ public class CustomerService {
                         () -> new CustomerException("User with id " + String.valueOf(id) + " does not exist")
                 );
 
+        if (!customer.getBookings().isEmpty()) {
+            throw new CustomerException("Cannot delete customer with active bookings");
+        }
+
         customerRepository.delete(customer);
     }
